Fixed remaining cells and moved PartSynchronizer into timed/
