+++ src/main/java/moe/nightfall/vic/integratedcircuits/cp/part/cell/PartBufferCell.java
@@ -15,8 +15,8 @@ public class PartBufferCell extends PartSimpleGate {
 	}
 
 	@Override
-	public void onInputChange(Vec2 pos, ICircuit parent, ForgeDirection side) {
-		super.onInputChange(pos, parent, side);
+	public void onInputChange(Vec2 pos, ICircuit parent) {
+		super.onInputChange(pos, parent);
 		notifyNeighbours(pos, parent);
 	}
